+++ src/entities/Notice/CustomNotice.tsx
@@ -1,4 +1,4 @@
-import Post from '@/entities/Post';
+import Post from '../Post/Post';
 
 /**
  *
@@ -13,7 +13,7 @@ const CustomNotice = ({ customNoticeList }: { customNoticeList: Notice[] }) => {
         <span className='text-[20px] font-bold leading-6 md:text-[28px]'>
           맞춤 공고
         </span>
-        <div className='inline-flex w-full items-start gap-1 overflow-x-scroll scrollbar-hide md:gap-[14px]'>
+        <div className='scrollbar-hide inline-flex w-full items-start gap-1 overflow-x-scroll md:gap-[14px]'>
           {customNoticeList.map(notice => {
             return <Post key={notice.item.id} noticeItem={notice.item} />;
           })}
