+++ src/entities/Post/DetailPost.tsx
