+++ src/pages/NoticeDetailPage/NoticeDetailPage.tsx
@@ -1,5 +1,5 @@
-import NoticeDetail from '@/widgets/NoticeDetail';
-import NoticeList from '@/widgets/NoticeList';
+import NoticeDetail from '@/entities/Notice/NoticeDetail';
+import NoticeList from '@/entities/Notice/NoticeList';
 
 // 임시로 만든 함수. '/public/data/mock.json' 에서 데이터를 가져옵니다.
 const getNotice = async () => {
